refactor(web): share marketing poster primitives (JOV-4882)
